creative-agent/pyproject.toml
[project]
name = "creative-agent"
version = "0.1.0"
description = "Creative Agent for generating marketing assets"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "pydantic>=2.9.0",
    "pydantic-settings>=2.6.0",
    "google-cloud-firestore>=2.19.0",
    "google-cloud-storage>=2.18.0",
    "google-cloud-pubsub>=2.26.0",
    "google-cloud-aiplatform>=1.71.0",
    "google-genai>=0.2.0",
    "google-adk>=1.0.0",
    "pillow>=10.4.0",
    "promote-autonomy-shared",
]

[dependency-groups]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=6.0.0",
    "httpx>=0.27.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"

[tool.coverage.run]
source = ["app"]
omit = ["tests/*"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
]

[tool.uv.sources]
promote-autonomy-shared = { path = "../shared", editable = true }
